fix(agents): keep bootstrap source identity private
